Extract block device code and fixup provision
